+++ SkiEngine/NCS/System/ISystem.cs
@@ -4,7 +4,7 @@ namespace SkiEngine.NCS.System
 {
     public interface ISystem
     {
-        void OnNodeZChanged(Node node, int previousZ);
+        void OnNodeZChanged(Node node, float previousZ);
         void OnComponentCreated(IComponent component);
         void OnComponentDestroyed(IComponent component);
     }
